Refactor using tail calls
In order to separate the parts and trying to use custom xdp_md meta data
values and tail calls this commit changes the structure from inlined
functions to separate XDP programs and combines them using tail calls.
Necessary additional data that is found for making the decision if to
reply with time exceeded messages is passed in the meta data section of
the XDP context, so it does not have to be fetched from the table again.