fix: allow tc layer 3 program tests
When testing layer 3 tc programs, the skb does not have the field
protocol set. This is an issue for bpf_skb_adjust _room which is used
for adjusting the packet's head for the additional headers.

This commit adds a fallback mechanisms and calls bpf_skb_change_head in
case skb->protocol is not set as required. This allows to run layer 3
tests now.